refactor + scale rooms
